+++ include/mxnet/c_api.h
@@ -392,18 +392,6 @@ MXNET_DLL int MXNDArraySlice(NDArrayHandle handle,
                              NDArrayHandle *out);
 
 /*!
- * \brief Slice the NDArray with non-default storage along axis 0.
- * \param handle the handle to the NDArray
- * \param slice_begin The beginning index of slice
- * \param slice_end The ending index of slice
- * \param out The NDArrayHandle of sliced NDArray
- * \return 0 when success, -1 when failure happens
- */
-MXNET_DLL int MXNDArraySliceEx(NDArrayHandle handle,
-                   mx_uint slice_begin,
-                   mx_uint slice_end,
-                   NDArrayHandle out);
-/*!
  * \brief Index the NDArray along axis 0.
  * \param handle the handle to the NDArray
  * \param idx the index
