+++ src/c_api/c_api.cc
@@ -320,16 +320,6 @@ int MXNDArraySlice(NDArrayHandle handle,
   API_END_HANDLE_ERROR(delete ptr);
 }
 
-int MXNDArraySliceEx(NDArrayHandle handle,
-                   mx_uint slice_begin,
-                   mx_uint slice_end,
-                   NDArrayHandle out) {
-  NDArray *ptr = static_cast<NDArray*>(out);
-  API_BEGIN();
-  static_cast<NDArray*>(handle)->SliceEx(slice_begin, slice_end, ptr);
-  API_END();
-}
-
 int MXNDArrayAt(NDArrayHandle handle,
                 mx_uint idx,
                 NDArrayHandle *out) {
