+++ src/ndarray/ndarray.cc
@@ -13,6 +13,7 @@
 #include <mshadow/tensor.h>
 #include "./ndarray_function.h"
 #include "../operator/tensor/matrix_op-inl.h"
+#include "../operator/tensor/init_op.h"
 #include "./autograd.h"
 
 #if MXNET_USE_OPENCV
@@ -62,52 +63,49 @@ NDArray NDArray::Reshape(const TShape &shape) const {
 NDArray NDArray::Slice(index_t begin, index_t end) const {
   using namespace autograd;
   using namespace mshadow;
-  NDArray ret = *this;
   CHECK(!is_none()) << "NDArray is not initialized";
   CHECK_GE(shape_[0], end) << "Slice end index out of range";
-  auto stype = storage_type();
-  CHECK_EQ(stype, kDefaultStorage);
-  size_t length = shape_.ProdShape(1, shape_.ndim());
-  MSHADOW_TYPE_SWITCH(ret.dtype(), DType, {
-    ret.byte_offset_ += begin * length * sizeof(DType);
-  });
-  ret.shape_[0] = end - begin;
-  if (AutogradRuntime::Get()->IsTraining()) {
-    // fake a slice_axis op
-    ret.entry_.clear();
-    const nnvm::Op* op = nnvm::Op::Get("slice_axis");
-    nnvm::NodeAttrs attrs;
-    attrs.op = op;
-    attrs.dict.insert({"axis", "0"});
-    attrs.dict.insert({"begin", std::to_string(begin)});
-    attrs.dict.insert({"end", std::to_string(end)});
-    op->attr_parser(&attrs);
-    std::vector<NDArray> inputs, outputs;
-    inputs.emplace_back(*this);
-    outputs.emplace_back(std::move(ret));
-    AutogradRuntime::Get()->RecordImperativeFCompute(
-      op, attrs, &inputs, &outputs);
-    return outputs[0];
-  } else {
-    return ret;
-  }
-}
-
-void NDArray::SliceEx(index_t begin, index_t end, NDArray *ret) const {
-  using namespace autograd;
-  using namespace mshadow;
-  CHECK(!is_none()) << "NDArray is not initialized";
-  CHECK_GE(shape_[0], end) << "Slice end index out of range";
-  auto stype = storage_type();
-  CHECK_NE(stype, kDefaultStorage);
-  if (stype == kCSRStorage) {
+  CHECK_NE(storage_type(), kUndefinedStorage);
+  if (storage_type() == kDefaultStorage) {
+    NDArray ret = *this;
+    auto stype = storage_type();
+    size_t length = shape_.ProdShape(1, shape_.ndim());
+    MSHADOW_TYPE_SWITCH(ret.dtype(), DType, {
+      ret.byte_offset_ += begin * length * sizeof(DType);
+    });
+    ret.shape_[0] = end - begin;
+    if (AutogradRuntime::Get()->IsTraining()) {
+      // fake a slice_axis op
+      ret.entry_.clear();
+      const nnvm::Op* op = nnvm::Op::Get("slice_axis");
+      nnvm::NodeAttrs attrs;
+      attrs.op = op;
+      attrs.dict.insert({"axis", "0"});
+      attrs.dict.insert({"begin", std::to_string(begin)});
+      attrs.dict.insert({"end", std::to_string(end)});
+      op->attr_parser(&attrs);
+      std::vector<NDArray> inputs, outputs;
+      inputs.emplace_back(*this);
+      outputs.emplace_back(std::move(ret));
+      AutogradRuntime::Get()->RecordImperativeFCompute(
+        op, attrs, &inputs, &outputs);
+      return outputs[0];
+    } else {
+      return ret;
+    }
+  } else if (storage_type() == kCSRStorage) {
+    // TODO(haibin) support auto_grad
+    TShape sliced_shape(Shape2(end-begin, shape()[1]));
     using namespace csr;
-    ret->shape_[0] = end - begin;
+    NDArray ret(storage_type(), TShape(Shape2(end-begin, shape()[1])),
+                ctx(), true, dtype_, ptr_->aux_types,
+                {TShape(Shape1(0)), TShape(Shape1(0))});
     NDArray src = *this;
     // destination NDArray shares the same variable
-    ret->ptr_->var = var();
+    ret.ptr_->var = var();
+
     Engine::Get()->PushSync([src, ret, begin, end](RunContext ctx) {
-      NDArray dst = *ret;
+      NDArray dst = ret;
       // create a new chunk for dst NDArray
       NDArray::Chunk chunk = *src.ptr_;
       // void indptr storage handle
@@ -147,10 +145,11 @@ void NDArray::SliceEx(index_t begin, index_t end, NDArray *ret) const {
       }
       }, ctx(), {}, {var()},
       FnProperty::kNormal, 0, PROFILER_MESSAGE_FUNCNAME);
+    return ret;
   } else {
-    LOG(FATAL) << "Slice not yet implemented for storage " << stype;
+    LOG(FATAL) << "Slice not yet implemented for storage " << storage_type();
   }
-  // TODO(haibin) support auto_grad for SliceEx
+  return NDArray();
 }
 
 NDArray NDArray::At(index_t idx) const {
@@ -380,6 +379,111 @@ void ScalarOp(const NDArray &lhs,
   }
 }
 
+size_t num_aux_data(NDArrayStorageType stype) {
+  size_t num = 0;
+  switch (stype) {
+    case kDefaultStorage: num = 0; break;
+    case kCSRStorage: num = 2; break;
+    case kRowSparseStorage: num = 1; break;
+     default: LOG(FATAL) << "Unknown storage type" << stype; break;
+  }
+  return num;
+}
+
+// Make a copy of a CSR NDArray
+template<typename from_xpu, typename to_xpu>
+inline void CopyFromToCsrImpl(const NDArray from, NDArray *to, RunContext ctx) {
+  using namespace mshadow;
+  CHECK_EQ(from.storage_type(), to->storage_type()) << "Copying with different storage type";
+  // if source storage is not initialized, fill destination with zeros
+  auto s = ctx.get_stream<to_xpu>();
+  if (!from.storage_initialized()) {
+    op::FillZerosCsrImpl<to_xpu>(s, to);
+    return;
+  }
+  // Allocate storage
+  to->CheckAndAllocAuxData(csr::kIndPtr, from.aux_shape(csr::kIndPtr));
+  to->CheckAndAllocAuxData(csr::kIdx, from.aux_shape(csr::kIdx));
+  to->CheckAndAllocData(from.aux_shape(csr::kIdx));
+  TBlob val = to->data();
+  TBlob indptr = to->aux_data(csr::kIndPtr);
+  TBlob idx = to->aux_data(csr::kIdx);
+  ndarray::Copy<from_xpu, to_xpu>(from.data(), &val,
+                                  from.ctx(), to->ctx(), ctx);
+  ndarray::Copy<from_xpu, to_xpu>(from.aux_data(csr::kIndPtr), &indptr,
+                                  from.ctx(), to->ctx(), ctx);
+  ndarray::Copy<from_xpu, to_xpu>(from.aux_data(csr::kIdx), &idx,
+                                  from.ctx(), to->ctx(), ctx);
+}
+
+// Make a copy of a row-sparse NDArray
+template<typename from_xpu, typename to_xpu>
+inline void CopyFromToRspImpl(const NDArray from, NDArray *to, RunContext ctx) {
+  using namespace mshadow;
+  CHECK_EQ(from.storage_type(), to->storage_type()) << "Copying with different storage type";
+  // if source is zeros, fill destination with zeros, too
+  auto s = ctx.get_stream<to_xpu>();
+  if (!from.storage_initialized()) {
+    op::FillZerosRspImpl<to_xpu>(s, to);
+    return;
+  }
+  auto aux_shape = from.aux_shape(rowsparse::kIdx);
+  to->CheckAndAlloc({aux_shape});
+  TBlob val = to->data();
+  TBlob idx = to->aux_data(rowsparse::kIdx);
+  ndarray::Copy<from_xpu, to_xpu>(from.data(), &val,
+                                  from.ctx(), to->ctx(), ctx);
+  ndarray::Copy<from_xpu, to_xpu>(from.aux_data(rowsparse::kIdx), &idx,
+                                  from.ctx(), to->ctx(), ctx);
+}
+
+// Make a copy of a dense NDArray
+template<typename from_xpu, typename to_xpu>
+inline void CopyFromToDnsImpl(const NDArray from, NDArray *to, RunContext ctx) {
+  using namespace mshadow;
+  CHECK_EQ(from.storage_type(), to->storage_type()) << "Copying with different storage type";
+  TBlob tmp = to->data();
+  ndarray::Copy<from_xpu, to_xpu>(from.data(), &tmp,
+                                  from.ctx(), to->ctx(), ctx);
+}
+
+// Make a copy of an NDArray based on storage type
+template<typename from_xpu, typename to_xpu>
+void CopyFromToImpl(const NDArray from, NDArray *to, RunContext ctx) {
+  using namespace std;
+  using namespace mshadow;
+  // if storage type doesn't match, cast the storage first
+  auto from_stype = from.storage_type();
+  auto to_stype = to->storage_type();
+  NDArray casted_nd;
+  if (from_stype != to_stype) {
+    TShape shape = from.shape();
+    auto from_ctx = from.ctx();
+    auto s = ctx.get_stream<from_xpu>();
+    // TODO(haibin) inplace conversion
+    if (to_stype == kDefaultStorage) {
+      casted_nd = NDArray(shape, from_ctx);
+    } else {
+      casted_nd = NDArray(to_stype, shape, from_ctx);
+    }
+    op::CastStorageComputeImpl<from_xpu>(s, from, casted_nd);
+  } else {
+    casted_nd = from;
+  }
+  if (to_stype == kDefaultStorage) {
+    CopyFromToDnsImpl<from_xpu, to_xpu>(casted_nd, to, ctx);
+  } else if (to_stype == kRowSparseStorage) {
+    CopyFromToRspImpl<from_xpu, to_xpu>(casted_nd, to, ctx);
+  } else if (to_stype == kCSRStorage) {
+    CopyFromToCsrImpl<from_xpu, to_xpu>(casted_nd, to, ctx);
+  } else {
+    LOG(FATAL) << "unknown storage type" << to_stype;
+  }
+  if (is_same<from_xpu, mshadow::gpu>::value || is_same<to_xpu, mshadow::gpu>::value) {
+    // Wait GPU kernel to complete
+    ctx.get_stream<gpu>()->Wait();
+  }
+}
 
 void CopyFromTo(const NDArray &from, NDArray *to, int priority) {
   if (from.var() == to->var()) {
@@ -406,6 +510,7 @@ void CopyFromTo(const NDArray &from, NDArray *to, int priority) {
       FnProperty::kNormal, priority, PROFILER_MESSAGE("CopyCPU2CPU"));
   } else {
 #if MXNET_USE_CUDA
+#if __CUDACC__
     if (a == cpu::kDevMask && b == gpu::kDevMask) {
       Engine::Get()->PushSync([from, ret](RunContext ctx) {
           NDArray nd(ret);
@@ -428,6 +533,7 @@ void CopyFromTo(const NDArray &from, NDArray *to, int priority) {
     } else {
       LOG(FATAL) << "unknown device mask";
     }
+#endif  // __CUDACC__
 #else
     LOG(FATAL) << MXNET_GPU_NOT_ENABLED_ERROR;
 #endif
@@ -707,9 +813,9 @@ void NDArray::Save(dmlc::Stream *strm) const {
   int32_t stype = storage_type();
   strm->Write(&stype, sizeof(stype));
 
-  const int32_t num_aux_data = NumAuxData(storage_type());
+  const int32_t nad = num_aux_data(storage_type());
   // save storage shape if ndarray is sparse
-  if (num_aux_data > 0) {
+  if (nad > 0) {
     storage_shape().Save(strm);
   }
 
@@ -737,8 +843,8 @@ void NDArray::Save(dmlc::Stream *strm) const {
   strm->Write(&type_flag, sizeof(type_flag));
 
   // save aux_types and aux_shapes
-  if (num_aux_data > 0) {
-    for (int i = 0; i < num_aux_data; ++i) {
+  if (nad > 0) {
+    for (int i = 0; i < nad; ++i) {
       int32_t aux_type_flag = aux_type(i);
       strm->Write(&aux_type_flag, sizeof(aux_type_flag));
       aux_shape(i).Save(strm);
@@ -753,8 +859,8 @@ void NDArray::Save(dmlc::Stream *strm) const {
   strm->Write(save_data.dptr_, type_size * save_data.shape_.Size());
 
   // save aux data
-  if (num_aux_data > 0) {
-    for (int i = 0; i < num_aux_data; ++i) {
+  if (nad > 0) {
+    for (int i = 0; i < nad; ++i) {
       TBlob save_data = nd_cpu.aux_data(i);
       // save aux_data
       CHECK(save_data.CheckContiguous());
@@ -821,11 +927,11 @@ bool NDArray::Load(dmlc::Stream *strm) {
   // load storage type
   int32_t stype;
   if (strm->Read(&stype, sizeof(stype)) != sizeof(stype)) return false;
-  const int32_t num_aux_data = NumAuxData(static_cast<NDArrayStorageType>(stype));
+  const int32_t nad = num_aux_data(static_cast<NDArrayStorageType>(stype));
 
   // load storage shape
   TShape sshape;
-  if (num_aux_data > 0) {
+  if (nad > 0) {
     if (!sshape.Load(strm)) return false;
   }
 
@@ -847,10 +953,10 @@ bool NDArray::Load(dmlc::Stream *strm) {
   // load aux_types and aux_shapes
   std::vector<int32_t> aux_types;
   std::vector<TShape> aux_shapes;
-  if (num_aux_data > 0) {
-    aux_types.resize(num_aux_data);
-    aux_shapes.resize(num_aux_data);
-    for (int i = 0; i < num_aux_data; ++i) {
+  if (nad > 0) {
+    aux_types.resize(nad);
+    aux_shapes.resize(nad);
+    for (int i = 0; i < nad; ++i) {
       // load aux_type(i)
       if (strm->Read(&aux_types[i], sizeof(aux_types[i])) != sizeof(aux_types[i])) return false;
       // load aux_shapes(i)
@@ -860,7 +966,7 @@ bool NDArray::Load(dmlc::Stream *strm) {
 
   // load data into CPU
   NDArray temp;
-  if (0 == num_aux_data) {
+  if (0 == nad) {
     temp = NDArray(shape, Context::CPU(), false, type_flag);
   } else {
     temp = NDArray(static_cast<NDArrayStorageType>(stype), shape,
@@ -874,8 +980,8 @@ bool NDArray::Load(dmlc::Stream *strm) {
   if (strm->Read(load_data.dptr_, nread) != nread) return false;
 
   // load aux_data
-  if (num_aux_data > 0) {
-    for (int i = 0; i < num_aux_data; ++i) {
+  if (nad > 0) {
+    for (int i = 0; i < nad; ++i) {
       load_data = temp.aux_data(i);
       type_size = mshadow::mshadow_sizeof(load_data.type_flag_);
       nread = type_size * load_data.Size();
