+++ src/operator/tensor/init_op.cu
@@ -10,7 +10,7 @@ namespace op {
 
 NNVM_REGISTER_OP(_zeros)
 .set_attr<FCompute>("FCompute<gpu>", FillCompute<gpu, 0>)
-.set_attr<FComputeEx>(FCOMP_EX_GPU, FillComputeZerosEx<gpu>);
+.set_attr<FComputeEx>("FComputeEx<gpu>", FillComputeZerosEx<gpu>);
 
 NNVM_REGISTER_OP(_ones)
 .set_attr<FCompute>("FCompute<gpu>", FillCompute<gpu, 1>);
