+++ src/operator/tensor/init_op.h
@@ -135,8 +135,8 @@ void FillZerosRspImpl(mshadow::Stream<xpu> *s, NDArray *dst) {
   // reset the shapes if it's not zeros
   auto storage_shape = dst->storage_shape();
   storage_shape[0] = 0;
-  dst->SetAuxShape(rowsparse::kIdx, TShape(mshadow::Shape1(0)));
-  dst->SetStorageShape(storage_shape);
+  dst->set_aux_shape(rowsparse::kIdx, TShape(mshadow::Shape1(0)));
+  dst->set_storage_shape(storage_shape);
 }
 
 // Fill a CSR NDArray with zeros by updating the aux shape.
@@ -145,9 +145,9 @@ void FillZerosCsrImpl(mshadow::Stream<xpu> *s, NDArray *dst) {
   if (!dst->storage_initialized()) return;
   // reset the shapes if it's not zeros
   TShape new_shape(mshadow::Shape1(0));
-  dst->SetAuxShape(csr::kIndPtr, new_shape);
-  dst->SetAuxShape(csr::kIdx, new_shape);
-  dst->SetStorageShape(new_shape);
+  dst->set_aux_shape(csr::kIndPtr, new_shape);
+  dst->set_aux_shape(csr::kIdx, new_shape);
+  dst->set_storage_shape(new_shape);
 }
 
 // This operator never needs to fall back, since there's no input NDArray
