+++ tests/python/unittest/test_sparse_ndarray.py
@@ -63,7 +63,7 @@ def test_sparse_nd_elementwise_fallback():
 def test_sparse_nd_zeros():
     def check_sparse_nd_zeros(stype, shape):
         zero = mx.nd.zeros(shape)
-        sparse_zero = mx.sparse_nd.zeros('row_sparse', shape)
+        sparse_zero = mx.nd.zeros(shape=shape, storage_type=stype)
         assert_almost_equal(sparse_zero.asnumpy(), zero.asnumpy())
 
     shape = rand_shape_2d()
@@ -135,7 +135,7 @@ def test_sparse_nd_basic():
 
 def test_sparse_nd_setitem():
     def check_sparse_nd_setitem(storage_type, shape, dst):
-        x = mx.sparse_nd.zeros(storage_type, shape)
+        x = mx.nd.zeros(shape=shape, storage_type=storage_type)
         x[:] = dst
         dst_nd = mx.nd.array(dst) if isinstance(dst, (np.ndarray, np.generic)) else dst
         assert same(x.asnumpy(), dst_nd.asnumpy())
@@ -167,7 +167,7 @@ def test_sparse_nd_slice():
 def test_sparse_nd_equal():
     for stype in ['row_sparse', 'csr']:
         shape = rand_shape_2d()
-        x = mx.sparse_nd.zeros(stype, shape)
+        x = mx.nd.zeros(shape=shape, storage_type=stype)
         y = sparse_nd_ones(shape, stype)
         z = x == y
         assert (z.asnumpy() == np.zeros(shape)).all()
@@ -178,7 +178,7 @@ def test_sparse_nd_equal():
 def test_sparse_nd_not_equal():
     for stype in ['row_sparse', 'csr']:
         shape = rand_shape_2d()
-        x = mx.sparse_nd.zeros(stype, shape)
+        x = mx.nd.zeros(shape=shape, storage_type=stype)
         y = sparse_nd_ones(shape, stype)
         z = x != y
         assert (z.asnumpy() == np.ones(shape)).all()
@@ -189,7 +189,7 @@ def test_sparse_nd_not_equal():
 def test_sparse_nd_greater():
     for stype in ['row_sparse', 'csr']:
         shape = rand_shape_2d()
-        x = mx.sparse_nd.zeros(stype, shape)
+        x = mx.nd.zeros(shape=shape, storage_type=stype)
         y = sparse_nd_ones(shape, stype)
         z = x > y
         assert (z.asnumpy() == np.zeros(shape)).all()
@@ -202,7 +202,7 @@ def test_sparse_nd_greater():
 def test_sparse_nd_greater_equal():
     for stype in ['row_sparse', 'csr']:
         shape = rand_shape_2d()
-        x = mx.sparse_nd.zeros(stype, shape)
+        x = mx.nd.zeros(shape=shape, storage_type=stype)
         y = sparse_nd_ones(shape, stype)
         z = x >= y
         assert (z.asnumpy() == np.zeros(shape)).all()
@@ -217,7 +217,7 @@ def test_sparse_nd_greater_equal():
 def test_sparse_nd_lesser():
     for stype in ['row_sparse', 'csr']:
         shape = rand_shape_2d()
-        x = mx.sparse_nd.zeros(stype, shape)
+        x = mx.nd.zeros(shape=shape, storage_type=stype)
         y = sparse_nd_ones(shape, stype)
         z = y < x
         assert (z.asnumpy() == np.zeros(shape)).all()
@@ -230,7 +230,7 @@ def test_sparse_nd_lesser():
 def test_sparse_nd_lesser_equal():
     for stype in ['row_sparse', 'csr']:
         shape = rand_shape_2d()
-        x = mx.sparse_nd.zeros(stype, shape)
+        x = mx.nd.zeros(shape=shape, storage_type=stype)
         y = sparse_nd_ones(shape, stype)
         z = y <= x
         assert (z.asnumpy() == np.zeros(shape)).all()
@@ -325,7 +325,7 @@ def test_sparse_nd_negate():
 
 def test_sparse_nd_output_fallback():
     shape = (10, 10)
-    out = mx.sparse_nd.zeros('row_sparse', shape)
+    out = mx.nd.zeros(shape=shape, storage_type='row_sparse')
     mx.nd.random_normal(shape=shape, out=out)
     assert(np.sum(out.asnumpy()) != 0)
 
@@ -333,7 +333,7 @@ def test_sparse_nd_output_fallback():
 def test_sparse_nd_astype():
     stypes = ['row_sparse', 'csr']
     for stype in stypes:
-        x = mx.sparse_nd.zeros(stype, rand_shape_2d(), dtype='float32')
+        x = mx.nd.zeros(shape=rand_shape_2d(), storage_type=stype, dtype='float32')
         y = x.astype('int32')
         assert(y.dtype == np.int32), y.dtype
 
@@ -359,48 +359,6 @@ def test_sparse_ndarray_pickle():
 
 
 def test_sparse_ndarray_save_load():
-    # TODO(junwu): This function is a duplicate of mx.nd.load
-    # which must be modified to use _ndarray_cls to generate
-    # dense/sparse ndarrays. However, a circular import issue
-    # arises when _ndarray_cls is used in mx.nd.load since
-    # ndarray.py and sparse_ndarray.py would import each other.
-    # We propose to put _ndarray_cls and all the functions calling
-    # it in ndarray.py and sparse_ndarray.py into a util file
-    # to resolve the circular import issue. This function will be
-    # kept till then.
-    def load(fname):
-        """Loads an array from file.
-        See more details in ``save``.
-        Parameters
-        ----------
-        fname : str
-            The filename.
-        Returns
-        -------
-        list of NDArray or dict of str to NDArray
-            Loaded data.
-        """
-        from mxnet.base import string_types, mx_uint, NDArrayHandle, check_call, c_str, _LIB
-        if not isinstance(fname, string_types):
-            raise TypeError('fname required to be a string')
-        out_size = mx_uint()
-        out_name_size = mx_uint()
-        import ctypes
-        handles = ctypes.POINTER(NDArrayHandle)()
-        names = ctypes.POINTER(ctypes.c_char_p)()
-        check_call(_LIB.MXNDArrayLoad(c_str(fname),
-                                      ctypes.byref(out_size),
-                                      ctypes.byref(handles),
-                                      ctypes.byref(out_name_size),
-                                      ctypes.byref(names)))
-        if out_name_size.value == 0:
-            return [_ndarray_cls(NDArrayHandle(handles[i])) for i in range(out_size.value)]
-        else:
-            assert out_name_size.value == out_size.value
-            from mxnet.base import py_str
-            return dict(
-                (py_str(names[i]), _ndarray_cls(NDArrayHandle(handles[i]))) for i in range(out_size.value))
-
     np.random.seed(0)
     repeat = 1
     stypes = ['default', 'row_sparse', 'csr']
@@ -418,14 +376,14 @@ def test_sparse_ndarray_save_load():
             assert isinstance(data_list1[-1], stype_dict[stype])
         mx.nd.save(fname, data_list1)
 
-        data_list2 = load(fname)
+        data_list2 = mx.nd.load(fname)
         assert len(data_list1) == len(data_list2)
         for x, y in zip(data_list1, data_list2):
             assert same(x.asnumpy(), y.asnumpy())
 
         data_map1 = {'ndarray xx %s' % i: x for i, x in enumerate(data_list1)}
         mx.nd.save(fname, data_map1)
-        data_map2 = load(fname)
+        data_map2 = mx.nd.load(fname)
         assert len(data_map1) == len(data_map2)
         for k, x in data_map1.items():
             y = data_map2[k]
